feat(interop): publish-ready interop crates; the gaze contract moves to vizij-arora-behavior
vizij-api-core, vizij-arora-host (1.0.0), and vizij-arora-behavior (1.0.0)
become publishable (license/repository metadata, version= on path deps), so
external hosts — the vizij-web standalone first — can run the interpreter and
the skill plane without this workspace. The portable look_at contract
(ids, parameters, signature, fragment parsing incl. the embedded override)
moves from the app's gaze module into vizij_arora_behavior::gaze; the app
keeps only the arora HostModule packaging. publish-crates.yml becomes a
working dispatch workflow: ordered, skip-if-published.
