+++ crates/interop/vizij-arora-host/Cargo.toml
@@ -1,12 +1,13 @@
 [package]
 name = "vizij-arora-host"
-version = "0.0.0"
+version = "1.0.0"
 edition.workspace = true
-publish = false
+license = "MIT OR Apache-2.0"
 description = "Portable Vizij host glue: compose the device behavior from graph sources, read the face bundle, pick the program to autoplay, and stage the neutral pose. Shared by the native app and vizij-arora-web (the browser runtime)."
+repository = "https://github.com/vizij-ai/vizij-rs"
 
 [dependencies]
-vizij-api-core = { path = "../../api/vizij-api-core" }
+vizij-api-core = { version = "1", path = "../../api/vizij-api-core" }
 anyhow = { workspace = true }
 serde_json = { workspace = true }
 log = "0.4"
